判断上期成绩是否为0

+++ csr-api/src/main/java/com/project/csr/service/StoreService.java
@@ -44,5 +44,5 @@ public interface StoreService extends IService<StorePo> {
      * @param parentId
      * @return
      */
-    List<StorePo> findByParentId(Integer parentId);
+    List<StorePo> findByParentId(Long parentId);
 }
